fastapi/etl/requirements.txt
pydantic==2.7.4
pydantic-settings==2.3.4
psycopg2-binary==2.9.9
elasticsearch==8.13.0
backoff==2.2.1
python-dotenv==1.0.1
